Assign threshold categories to Enterococcus data

Assign threshold categories to Enterococcus data

Usage

anlz_enteromap(
  fibdata,
  yrsel = NULL,
  mosel = NULL,
  wetdry = FALSE,
  precipdata = NULL,
  temporal_window = NULL,
  wet_threshold = NULL
)

Arguments

fibdata

data frame of Enterococcus sample data as returned by enterodata or anlz_fibwetdry

yrsel

optional numeric to filter data by year

mosel

optional numeric to filter data by month

wetdry

logical; if TRUE, incorporate wet/dry differences (this will result in a call to anlz_fibwetdry, in which case temporal_window and wet_threshold are required). if FALSE (default), do not differentiate between wet and dry samples.

precipdata

input data frame as returned by read_importrain. columns should be: station, date (yyyy-mm-dd), rain (in inches). The object catch_precip has this data from 1995-2023 for select Enterococcus stations. If NULL, defaults to catch_precip.

temporal_window

numeric; required if wetdry is TRUE. number of days precipitation should be summed over (1 = day of sample only; 2 = day of sample + day before; etc.)

wet_threshold

numeric; required if wetdry is TRUE. inches accumulated through the defined temporal window, above which a sample should be defined as being from a 'wet' time period

Value

A data.frame similar to fibdata with additional columns describing station categories and optionally filtered by arguments passed to the function

Details

This function is based on anlz_fibmap, but is specific to Enterococcus data downloaded via read_importentero. It creates categories for mapping using show_enteromap. Optionally, if samples have been defined as 'wet' or not via anlz_fibwetdry, this can be represented via symbols on the map. Categories based on relevant thresholds are assigned to each observation. The categories are specific to Enterococcus in marine waters (class of 2 or 3M). A station is categorized into one of four ranges defined by the thresholds as noted in the cat column of the output, with corresponding colors appropriate for each range as noted in the col column of the output.

Examples

anlz_enteromap(enterodata, yrsel = 2020, mosel = 9)
#>                  station   yr mo Latitude Longitude ecocci       cat     col
#> 1 21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35 #2DC938
#> 2 21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35 #2DC938
#> 3 21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35 #2DC938
#> 4       21FLHILL_WQX-101 2020  9 28.02571 -82.63164    380 130 - 999 #EE7600
#> 5       21FLHILL_WQX-102 2020  9 28.01060 -82.60780    380 130 - 999 #EE7600
#> 6     21FLPDEM_WQX-19-02 2020  9 27.92840 -82.74575     10      < 35 #2DC938
#>            ind  indnm conc
#> 1 Enterococcus ecocci   10
#> 2 Enterococcus ecocci   10
#> 3 Enterococcus ecocci   10
#> 4 Enterococcus ecocci  380
#> 5 Enterococcus ecocci  380
#> 6 Enterococcus ecocci   10

# differentiate wet/dry samples in that time frame
anlz_enteromap(enterodata, yrsel = 2020, mosel = 9, wetdry = TRUE,
               temporal_window = 2, wet_threshold = 0.5)
#>                  station   yr mo Latitude Longitude ecocci       cat     col
#> 1 21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35 #2DC938
#> 2 21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35 #2DC938
#> 3 21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35 #2DC938
#> 4       21FLHILL_WQX-101 2020  9 28.02571 -82.63164    380 130 - 999 #EE7600
#> 5       21FLHILL_WQX-102 2020  9 28.01060 -82.60780    380 130 - 999 #EE7600
#> 6     21FLPDEM_WQX-19-02 2020  9 27.92840 -82.74575     10      < 35 #2DC938
#>            ind  indnm conc wet_sample
#> 1 Enterococcus ecocci   10      FALSE
#> 2 Enterococcus ecocci   10      FALSE
#> 3 Enterococcus ecocci   10      FALSE
#> 4 Enterococcus ecocci  380      FALSE
#> 5 Enterococcus ecocci  380       TRUE
#> 6 Enterococcus ecocci   10      FALSE