Assign threshold categories to Fecal Indicator Bacteria (FIB) data

Usage

anlz_fibcat(fibdata, yrsel = NULL, mosel = NULL, areasel = NULL)

Arguments

fibdata

input FIB data.frame as returned by read_importfib

yrsel

optional numeric value to filter output by years in fibdata

mosel

optional numeric value to filter output by month in fibdata

areasel

optional character string to filter output by stations in the "Alafia" or "Hillsborough} River basins, see details} } { A \code{data.frame} similar to \code{fibdata} with additional columns describing station categories and optionally filtered by arguments passed to the function } \description{ Assign threshold categories to Fecal Indicator Bacteria (FIB) data } \details{ Categories based on relevant thresholds are assigned to each observation. The categories are specific to E. coli or Enterococcus and are assigned based on the station class as freshwater (\code{class} as 1 or 3F) or marine (\code{class} as 2 or 3M), respectively. A station is categorized into one of four ranges defined by the thresholds as noted in the \code{cat} column of the output, with corresponding colors appropriate for each range as noted in the \code{col} column of the output.The \code{areasel} argument can indicate either \code{"Alafia"} or \code{"Hillsborough for the corresponding river basins, where rows in fibdata are filtered based on the the selection. The Alafia River basin includes values in the area column of fibdata as "Alafia River" and "Alafia River Tributary". The Hillsborough River basin includes values in the area column of fibdat as "Hillsborough River", "Hillsborough River Tributary", "Lake Thonotosassa", "Lake Thonotosassa Tributary", and "Lake Roberta". Not all areas may be present based on the selection.

# assign categories to all anlz_fibcat(fibdata)# filter by year, month, and area anlz_fibcat(fibdata, yrsel = 2020, mosel = 7, areasel = 'Alafia')

anlz