Identify Fecal Indicator Bacteria samples as coming from a 'wet' or 'dry' time period

Identify Fecal Indicator Bacteria samples as coming from a 'wet' or 'dry' time period

Usage

anlz_fibwetdry(fibdata, precipdata, temporal_window = 2, wet_threshold = 0.5)

Arguments

fibdata

input data frame

precipdata

input data frame as returned by read_importrain. columns should be: station, date (yyyy-mm-dd), rain (in inches). The object catch_precip has this data from 1995-2023 for select stations.

temporal_window

numeric, number of days precipitation should be summed over (1 = day of sample only; 2 = day of sample + day before; etc.)

wet_threshold

numeric, inches accumulated through the defined temporal window, above which a sample should be defined as being from a 'wet' time period

Value

a data frame; the original fibdata data frame with three additional columns. rain_sampleDay is the total rain (inches) on the day of sampling, rain_total is the total rain (inches) for the period of time defined by temporal_window, and wet_sample is logical, indicating whether the rainfall for that station's catchment exceeded the amount over the time period specified in args.

Details

This function allows the user to specify a threshold for declaring a sample to be taken after an important amount of rain over an important amount of days, and declaring it to be 'wet'. This is of interest because samples taken after significant precipitation (definitions of this vary, which is why the user can specify desired thresholds) are more likely to exceed relevant bacterial thresholds. Identifying samples as 'wet' or not allows for calculation of further indices for wet and dry subsets of samples.

Examples

entero_wetdry <- anlz_fibwetdry(enterodata, catch_precip)
head(entero_wetdry)
#>         date                station ecocci ecocci_censored Latitude Longitude
#> 1 2019-01-07 21FLDOH_WQX-MANATEE152     10            TRUE 27.53593 -82.73678
#> 2 2019-01-22 21FLDOH_WQX-MANATEE152     10            TRUE 27.53593 -82.73678
#> 3 2019-02-05 21FLDOH_WQX-MANATEE152     20           FALSE 27.53593 -82.73678
#> 4 2019-02-18 21FLDOH_WQX-MANATEE152     10            TRUE 27.53593 -82.73678
#> 5 2019-03-05 21FLDOH_WQX-MANATEE152     10            TRUE 27.53593 -82.73678
#> 6 2019-03-18 21FLDOH_WQX-MANATEE152     10            TRUE 27.53593 -82.73678
#>   ecocci_units     time time_zone MDL   yr mo rain_sampleDay rain_total
#> 1    MPN/100mL 10:50:00       EST  10 2019  1     0.00000000  0.0000000
#> 2    MPN/100mL 11:10:00       EST  10 2019  1     0.00000000  0.0000000
#> 3    MPN/100mL 10:40:00       EST  10 2019  2     0.00000000  0.0000000
#> 4    MPN/100mL 11:25:00       EST  10 2019  2     0.00000000  0.0000000
#> 5    MPN/100mL 10:50:00       EST  10 2019  3     0.06957143  0.1541429
#> 6    MPN/100mL 10:40:00       EDT  10 2019  3     0.20728571  0.2124286
#>   wet_sample
#> 1      FALSE
#> 2      FALSE
#> 3      FALSE
#> 4      FALSE
#> 5      FALSE
#> 6      FALSE