Get Tampa Bay Benthic Index scores

Get Tampa Bay Benthic Index scores

anlz_tbbiscr(benthicdata)

Arguments

benthicdata
nested tibble formatted from read_importbenthic

Value

A single data frame of TBBI scores for each site.

Details

This function calculates scores for the TBBI based on station, taxa, and field sample data. The total TBBI scores are returned as TBBI and TBBICat, where the latter is a categorical description of the scores.

See also

Examples

anlz_tbbiscr(benthicdata)

#> # A tibble: 4,516 x 14
#>    StationID AreaAbbr FundingProject ProgramID ProgramName    Latitude Longitude
#>        <int> <chr>    <chr>              <int> <chr>             <dbl>     <dbl>
#>  1       448 BB       Apollo Beach           4 Benthic Monit…     27.8     -82.4
#>  2       449 BB       Apollo Beach           4 Benthic Monit…     27.8     -82.4
#>  3       450 BB       Apollo Beach           4 Benthic Monit…     27.8     -82.4
#>  4       451 BB       Apollo Beach           4 Benthic Monit…     27.8     -82.4
#>  5       452 BB       Apollo Beach           4 Benthic Monit…     27.8     -82.4
#>  6       453 BB       Apollo Beach           4 Benthic Monit…     27.8     -82.4
#>  7       454 BB       Apollo Beach           4 Benthic Monit…     27.8     -82.4
#>  8       455 BB       Apollo Beach           4 Benthic Monit…     27.8     -82.4
#>  9       456 BB       Apollo Beach           4 Benthic Monit…     27.8     -82.4
#> 10       457 BB       Apollo Beach           4 Benthic Monit…     27.8     -82.4
#> # … with 4,506 more rows, and 7 more variables: date <date>, yr <dbl>,
#> #   TotalAbundance <dbl>, SpeciesRichness <dbl>, TBBI <dbl>, TBBICat <chr>,
#> #   Salinity <dbl>