Estimate tidal creek report card scores

Estimate tidal creek report card scores

anlz_tdlcrk(tidalcreeks, iwrraw, tidtrgs = NULL, yr = 2018)

Arguments

tidalcreeks
sf object for population of tidal creeks

iwrraw
FDEP impaired waters rule run 56 data base as data.frame

tidtrgs
optional data.frame for tidal creek nitroge targets, defaults to tidaltargets

yr
numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date

Value

A data.frame with the report card scores for each creek, as red, orange, yellow, green, or no data

See also

Examples

anlz_tdlcrk(tidalcreeks, iwrraw, yr = 2018)

#> # A tibble: 581 x 9
#>       id wbid  JEI   class target caution action threshold score  
#>    <int> <chr> <chr> <chr>  <int>   <int>  <int>     <int> <chr>  
#>  1     1 2052  CC01  3M        NA      NA     NA        NA No Data
#>  2     2 1983B CC01  2         NA      NA     NA        NA No Data
#>  3     3 1983B CC02  2          1      NA     NA        NA Green  
#>  4     4 2067  CC02  3M         1      NA     NA        NA Green  
#>  5     5 1983B CC03  2         NA      NA     NA        NA No Data
#>  6     6 2068  CC03  3M        NA      NA     NA        NA No Data
#>  7     7 2078A CC04  2          2      NA     NA        NA Green  
#>  8     8 2078B CC05  2          2      NA     NA        NA Green  
#>  9     9 2078A CC05  2         NA      NA     NA        NA No Data
#> 10    10 2065C CC06  2         NA      NA     NA        NA No Data
#> # … with 571 more rows