Analyze tidal creek water quality indicators

Estimate tidal creek water quality indicators to support report card scores

anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2021, radar = FALSE)

Arguments

tidalcreeks
sf object for population of tidal creeks

iwrraw
FDEP impaired waters rule data base as data.frame

yr
numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date

radar
logical indicating if output is for show_tdlcrkradar, see details

Value

A data.frame with the indicator values for each tidal creek

Details

Annual geometric means for additional water quality data available at each wbid, JEI combination. Florida trophic state index values are also estimated where data are available.

Nitrogen ratios are estimated for JEIs that cover source (upstream, freshwater) and tidal (downstream) WBIDs, defined as the ratio of concentrations between the two (i.e., ratios > 1 mean source has higher concentrations). Nitrogen ratios for a given year reflect the ratio of the median nitrogen concentrations when they were measured in both a source and tidal segment during the same day. Note that a ratio of one can be obtained if both the source and tidal segments are at minimum detection.

Indicators for years where more than 10% of observations exceed DO saturation criteria are also estimated. The do_bnml and do_prop columns show a 1 or 0 for a given year to indicate if more than ten percent of observations were below DO percent saturation of 42. The first column is based on a binomial probability exceedance that takes into account the number of observations in the year and the second column is based on a simple proportional estimate from the raw data.

If radar = TRUE, output is returned in a format for use with show_tdlcrkradar Specifically, results are calculated as the percentage of years where an indicator exceeds a relevant threshold. This only applies to the marine WBIDs of the tidal creeks (Florida DEP class 2, 3M). Six indicators are returned with percentage exceedances based on total nitrogen (tn_ind) greater than 1.1 mg/L, chlorophyll (chla_ind) greater than 11 ug/L, trophic state index (tsi_ind) greater than 55 (out of 100), nitrate/nitrite ratio between marine and upstream segments (nox_ind) greater than one, chlorophyll and total nitrogen ratios > 15, and percentage of years more where than ten percent of observations were below DO percent saturation of 42.

Examples

dat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2021)
head(dat)
#>   id       name  JEI wbid class year    CHLAC     COLOR COND       DO    DOSAT
#> 1  1 Rock Creek CC01 2052    3M 2011 1.394022  3.399019   NA 4.249811 58.26457
#> 2  1 Rock Creek CC01 2052    3M 2012 2.281227 15.159408   NA 2.800000 48.48755
#> 3  1 Rock Creek CC01 2052    3M 2013 2.725391  9.607368   NA 6.819091 73.50770
#> 4  1 Rock Creek CC01 2052    3M 2014 2.380249 14.694735   NA 3.308289 60.31523
#> 5  1 Rock Creek CC01 2052    3M 2015 2.647817 21.911128   NA 4.034172 55.41609
#> 6  1 Rock Creek CC01 2052    3M 2016 2.265462 19.130797   NA 3.980251 55.96637
#>          NO23 ORGN    SALIN       TKN        TN         TP TSS     TURB
#> 1 0.006467844   NA 33.67785 0.4927774 0.4981428 0.05831404  NA 2.532379
#> 2 0.009742973   NA 26.49843 0.6707901 0.6803095 0.06762831  NA 2.085028
#> 3 0.010611775   NA 30.90561 0.6510575 0.7108146 0.06842811  NA 1.698004
#> 4 0.008132924   NA 26.63714 0.6161702 0.6228782 0.06938150  NA 2.044623
#> 5 0.009361504   NA 19.10494 0.6465161 0.5678241 0.07602770  NA 1.814818
#> 6 0.008532522   NA 25.84531 0.5813818 0.5408161 0.08330509  NA 2.082253
#>   chla_tn_ratio tn_tp_ratio chla_tsi   tn_tsi  tn2_tsi   tp_tsi  tp2_tsi
#> 1      2.798439    8.542416 21.58358 42.20200 44.62164 57.22468 72.15389
#> 2      3.353220   10.059537 28.67587 48.37289 51.32120 59.98090 75.65103
#> 3      3.834180   10.387756 31.23761 49.24140 52.26414 60.19958 75.92850
#> 4      3.821371    8.977584 29.28775 46.62660 49.42526 60.45694 76.25504
#> 5      4.663093    7.468648 30.82179 44.79432 47.43600 62.15842 78.41391
#> 6      4.188970    6.491994 28.57601 43.82942 46.38844 63.85868 80.57123
#>    nut_tsi      tsi no23_source no23_tidal no23_ratio do_bnml do_prop
#> 1 44.62164 33.10261          NA         NA         NA       0       0
#> 2 54.17690 41.42638          NA         NA         NA       1       1
#> 3 54.72049 42.97905          NA         NA         NA       0       0
#> 4 49.42526 39.35651          NA         NA         NA       0       1
#> 5 47.43600 39.12890          NA         NA         NA       0       1
#> 6 46.38844 37.48223          NA         NA         NA       0       1