Get attainment categories for a chosen year

Get attainment categories for a chosen year

anlz_yrattain(epcdata, yrsel)

Arguments

epcdata
data frame of epc data returned by read_importwq

yrsel
numeric indicating chosen year

Value

A data.frame for the chosen year and all bay segments showing the bay segment averages for chloropyll concentration, light attenuations, segment targets, and attainment categories.

See also

Examples

# defaults to current year
anlz_yrattain(epcdata, yrsel = 2018)

#> # A tibble: 4 x 6
#>   bay_segment chla_val chla_target la_val la_target outcome
#>   <fct>          <dbl>       <dbl>  <dbl>     <dbl> <chr>  
#> 1 OTB             9.22         8.5  0.678      0.83 yellow 
#> 2 HB             13.9         13.2  1.09       1.58 green  
#> 3 MTB             7.05         7.4  0.570      0.83 green  
#> 4 LTB             4.65         4.6  0.593      0.63 green