Download latest water quality file from epchc.org

Download latest water quality file from epchc.org

get_epchc_wq(xlsx, download_latest_epchc = TRUE)

Arguments

xlsx
chr string path for local excel file, to overwrite it not current

download_latest_epchc
logical to download latest file regardless of local copy

Value

The local copy specified in the path by xlsx is overwritten by the new file is not current or download_latest_epchc = TRUE. The function does nothing if download_latest_epchc = FALSE.

Details

The local copy is checked against a temporary file downloade from ftp://ftp.epchc.org/EPC_ERM_FTP/WQM_Reports/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx. The local file is replaced with the downloaded file if the MD5 hashes are different.

Examples

# NOT RUN {
test_xlsx <- file.path(dir_gdrive, "data/wq/2018_Results_Updated.xls")
get_epch_wq(test_xlsx)
# }