Pasco County FIB data as of 20250304

Pasco County FIB data as of 20250304

Usage

pascofibdata

Format

A data frame with 209 rows and 13 variables:

pasco_station

chr, Station name

SampleTime

POSIXct, Date/time of sampling

class

chr, Waterbody class (Fresh, Marine)

yr

num, Year of sampling

mo

num, Month of sampling

Latitude

num, Latitude, WGS84

Longitude

num, Latitude, WGS84

Sample_Depth_m

num, Depth of sample, meters

var

chr, Variable name (ecoli, entero

val

num, Value of variable

uni

num, Units of variable

qual

num, Qualifier code

area

chr, Location name

Details

This dataset includes FIB data from Pasco County (21FLPASC_WQX) where station class is marine (estuary) and Enterococcus data is present or the station class is freshwater (stream/river, reservoir) and E. coli data is present. The data is formatted from the raw data loaded from read_importwqp.

Examples

if (FALSE) { # \dontrun{
pascofibdata <- read_importwqp('21FLPASC_WQX', type = 'fib')

nrow(pascofibdata)
ncol(pascofibdata)

save(pascofibdata, file = 'data/pascofibdata.RData')
} # }