Format benthic data for the Tampa Bay Benthic Index

Format benthic data for the Tampa Bay Benthic Index

read_formbenthic(pathin)

Arguments

pathin

A path to unzipped csv files with base tables used to calculate benthic index

Value

A nested tibble of station, field sample, and taxa data

Details

Function is used internally within read_importbenthic

Examples

if (FALSE) {

# location to download data
path <- '~/Desktop/benthic.zip'

# download
urlin <- 'ftp://ftp.epchc.org/EPC_ERM_FTP/Benthic_Monitoring/DataDeliverables/BaseTables.zip'
read_dlcurrent(path, download_latest = TRUE, urlin = urlin)

# unzip
tmppth <- tempfile()
utils::unzip(path, exdir = tmppth, overwrite = TRUE)

# format benthic data
read_formbenthic(pathin = tmppth)

# remove temporary path
unlink(tmppth, recursive = TRUE)

}