Download Enterococcus data from the Water Quality Portal

Download Enterococcus data from the Water Quality Portal

Usage

read_importentero(stas = NULL, startDate, endDate)

Arguments

stas

character, a vector of stations. If NULL, defaults to all stations in catchprecip.

startDate

character, starting date of observations as YYYY-MM-DD

endDate

character, ending date of observations as YYYY-MM-DD

Value

a data frame containing one row for each sample. Columns returned are:

date

date, sample date

station

character, sample station

ecocci

numeric, Enterococcus concentration

ecocci_censored

logical, whether ecocci value was below the laboratory MDL, minimum detection limit

ecocci_units

character, units of measurement for ecocci

qualifier

qualifier codes associated with sample

LabComments

lab comments on sample

Latitude

numeric, latitude in decimal degrees

Longitude

numeric, longitude in decimal degrees

time

character, sample time

time_zone

character, sample time zone

MDL

numeric, minimum detection limit for the lab and time the sample was analyzed

yr

numeric, year of sample date

mo

numeric, month of sample date

Details

Retrieves Enterococcus sample data from selected stations and date range from the Water Quality Portal, https://www.waterqualitydata.us

Examples

if (FALSE) { # \dontrun{
# stations to download
stations <- c('21FLHILL_WQX-101',
'21FLHILL_WQX-102',
'21FLHILL_WQX-103')

# download and read the data
entero_in <- read_importentero(stas = stations, startDate = '2023-01-01', endDate = '2023-02-01')

head(entero_in)

} # }