Load local phytoplankton cell count file

Load local phytoplankton cell count file

read_importphyto(xlsx, download_latest = FALSE, na = "")

Arguments

xlsx
chr string path for local excel file, to overwrite if not current

download_latest
logical passed to read_dlcurrent to download raw data and compare with existing in xlsx if available

na
chr vector of strings to interpret as NA, passed to read_xlsx

Value

A data.frame of formatted water quality data.

Details

Pytoplankton cell count data downloaded from ftp://ftp.epchc.org/EPC_ERM_FTP/WQM_Reports/, file PlanktonDataList_ThroughCurrentReportMonth.xlsx

See also

Examples

if (FALSE) {
# file path
xlsx <- '~/Desktop/phyto_data.xlsx'

# load and assign to object
phytodata <- read_importphyto(xlsx)
}