Creates a plotly matrix from any matrix function input

Creates a plotly matrix from any matrix function input

show_matrixplotly(
  mat,
  family = NA,
  tooltip = "Result",
  width = NULL,
  height = NULL
)

Arguments

mat

input matrix as output from show_matrix, show_segmatrix, show_wqmatrix, or show_tbnimatrix

family

optional chr string indicating font family for text labels

tooltip

chr string indicating the column name for tooltip

width

numeric for width of the plot in pixels

height

numeric for height of the plot in pixels

Value

A plotly data object

Examples

mat <- show_wqmatrix(epcdata)
show_matrixplotly(mat)