Plot chlorophyll and secchi data together with matrix outcomes

Plot chlorophyll and secchi data together with matrix outcomes

Usage

show_segplotly(
  epcdata,
  bay_segment = c("OTB", "HB", "MTB", "LTB"),
  yrrng = c(1975, 2021),
  family = NULL,
  partialyr = FALSE,
  width = NULL,
  height = NULL
)

Arguments

epcdata

data frame of epc data returned by read_importwq

bay_segment

chr string for the bay segment, one of "OTB", "HB", "MTB", "LTB"

yrrng

numeric for year range to plot

family

optional chr string indicating font family for text labels

partialyr

logical indicating if incomplete annual data for the most recent year are approximated by five year monthly averages for each parameter

width

numeric for width of the plot in pixels

height

numeric for height of the plot in pixels

Value

An interactive plotly object

Details

This function combines outputs from show_thrplot and show_segmatrix for a selected bay segment. The plot is interactive and can be zoomed by dragging the mouse pointer over a section of the plot. Information about each cell or value can be seen by hovering over a location in the plot.

Examples

show_segplotly(epcdata)