Assign threshold categories to Enterococcus data

Assign threshold categories to Enterococcus data

Usage

anlz_enteromap(
  fibdata,
  yrsel = NULL,
  mosel = NULL,
  wetdry = FALSE,
  precipdata = NULL,
  temporal_window = NULL,
  wet_threshold = NULL
)

Arguments

fibdata

data frame of Enterococcus sample data as returned by enterodata or anlz_fibwetdry

yrsel

optional numeric to filter data by year

mosel

optional numeric to filter data by month

wetdry

logical; if TRUE, incorporate wet/dry differences (this will result in a call to anlz_fibwetdry, in which case temporal_window and wet_threshold are required). If FALSE (default), do not differentiate between wet and dry samples.

precipdata

input data frame as returned by read_importrain. columns should be: station, date (yyyy-mm-dd), rain (in inches). The object catchprecip has this data from 1995-2023 for select Enterococcus stations. If NULL, defaults to catchprecip.

temporal_window

numeric; required if wetdry is TRUE. number of days precipitation should be summed over (1 = day of sample only; 2 = day of sample + day before; etc.)

wet_threshold

numeric; required if wetdry is TRUE. inches accumulated through the defined temporal window, above which a sample should be defined as being from a 'wet' time period

Value

A data.frame similar to fibdata with additional columns describing station categories and optionally filtered by arguments passed to the function

Details

This function is based on anlz_fibmap, but is specific to Enterococcus data downloaded via read_importentero. It creates categories for mapping using show_enteromap. Optionally, if samples have been defined as 'wet' or not via anlz_fibwetdry, this can be represented via symbols on the map. Categories based on relevant thresholds are assigned to each observation. The categories are specific to Enterococcus in marine waters (class of 2 or 3M). A station is categorized into one of four ranges defined by the thresholds as noted in the cat column of the output, with corresponding colors appropriate for each range as noted in the col column of the output.

Examples

anlz_enteromap(enterodata, yrsel = 2020, mosel = 9)
#>                       station   yr mo Latitude Longitude ecocci       cat
#> 1          21FLCOSP_WQX-32-03 2020  9 27.80869 -82.61622     20      < 35
#> 2          21FLCOSP_WQX-44-02 2020  9 27.79258 -82.62640     41  35 - 129
#> 3          21FLCOSP_WQX-48-03 2020  9 27.71173 -82.68386     31      < 35
#> 4  21FLCOSP_WQX-CENTRAL CANAL 2020  9 27.72590 -82.68354     10      < 35
#> 5        21FLCOSP_WQX-COSP580 2020  9 27.73746 -82.69376     41  35 - 129
#> 6    21FLCOSP_WQX-NORTH CANAL 2020  9 27.72983 -82.68853     41  35 - 129
#> 7          21FLCOSP_WQX-SC-01 2020  9 27.75198 -82.63928    480 130 - 999
#> 8    21FLCOSP_WQX-SOUTH CANAL 2020  9 27.72461 -82.68794    360 130 - 999
#> 9      21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35
#> 10     21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35
#> 11     21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35
#> 12           21FLHILL_WQX-101 2020  9 28.02571 -82.63164    380 130 - 999
#> 13           21FLHILL_WQX-102 2020  9 28.01060 -82.60780    380 130 - 999
#> 14           21FLHILL_WQX-103 2020  9 27.99741 -82.58651    490 130 - 999
#> 15           21FLHILL_WQX-104 2020  9 27.98187 -82.56260    725 130 - 999
#> 16           21FLHILL_WQX-109 2020  9 27.94581 -82.40167      1      < 35
#> 17           21FLHILL_WQX-112 2020  9 27.70443 -82.44828    220 130 - 999
#> 18           21FLHILL_WQX-133 2020  9 27.91501 -82.40210   3100     > 999
#> 19           21FLHILL_WQX-144 2020  9 27.83802 -82.38191    620 130 - 999
#> 20           21FLHILL_WQX-180 2020  9 27.71599 -82.47044     56  35 - 129
#> 21             21FLHILL_WQX-2 2020  9 27.94183 -82.45854    280 130 - 999
#> 22            21FLHILL_WQX-22 2020  9 27.60810 -82.57120      2      < 35
#> 23           21FLHILL_WQX-264 2020  9 27.68253 -82.49677     40  35 - 129
#> 24           21FLHILL_WQX-560 2020  9 27.90522 -82.40163    600 130 - 999
#> 25           21FLHILL_WQX-561 2020  9 27.93162 -82.39926   3500     > 999
#> 26           21FLHILL_WQX-592 2020  9 27.77489 -82.38808    320 130 - 999
#> 27           21FLHILL_WQX-594 2020  9 27.81923 -82.38252    300 130 - 999
#> 28           21FLHILL_WQX-595 2020  9 27.82610 -82.38168   6100     > 999
#> 29           21FLHILL_WQX-596 2020  9 27.87821 -82.39036    580 130 - 999
#> 30           21FLHILL_WQX-597 2020  9 27.88509 -82.39170   1350     > 999
#> 31           21FLHILL_WQX-598 2020  9 27.89004 -82.39709    525 130 - 999
#> 32            21FLHILL_WQX-74 2020  9 27.85909 -82.38334     85  35 - 129
#> 33         21FLPDEM_WQX-12-02 2020  9 28.02048 -82.68955    703 130 - 999
#> 34         21FLPDEM_WQX-19-02 2020  9 27.92840 -82.74575     10      < 35
#> 35         21FLPDEM_WQX-22-01 2020  9 27.92346 -82.71765   1990     > 999
#> 36         21FLPDEM_WQX-24-02 2020  9 27.91209 -82.70025    171 130 - 999
#> 37         21FLPDEM_WQX-39-01 2020  9 27.75798 -82.73151    717 130 - 999
#> 38         21FLPDEM_WQX-39-05 2020  9 27.76021 -82.72979    727 130 - 999
#> 39       21FLTPA_WQX-G1SW0099 2020  9 27.86294 -82.61633    908 130 - 999
#>        col          ind  indnm conc
#> 1  #2DC938 Enterococcus ecocci   20
#> 2  #E9C318 Enterococcus ecocci   41
#> 3  #2DC938 Enterococcus ecocci   31
#> 4  #2DC938 Enterococcus ecocci   10
#> 5  #E9C318 Enterococcus ecocci   41
#> 6  #E9C318 Enterococcus ecocci   41
#> 7  #EE7600 Enterococcus ecocci  480
#> 8  #EE7600 Enterococcus ecocci  360
#> 9  #2DC938 Enterococcus ecocci   10
#> 10 #2DC938 Enterococcus ecocci   10
#> 11 #2DC938 Enterococcus ecocci   10
#> 12 #EE7600 Enterococcus ecocci  380
#> 13 #EE7600 Enterococcus ecocci  380
#> 14 #EE7600 Enterococcus ecocci  490
#> 15 #EE7600 Enterococcus ecocci  725
#> 16 #2DC938 Enterococcus ecocci    1
#> 17 #EE7600 Enterococcus ecocci  220
#> 18 #CC3231 Enterococcus ecocci 3100
#> 19 #EE7600 Enterococcus ecocci  620
#> 20 #E9C318 Enterococcus ecocci   56
#> 21 #EE7600 Enterococcus ecocci  280
#> 22 #2DC938 Enterococcus ecocci    2
#> 23 #E9C318 Enterococcus ecocci   40
#> 24 #EE7600 Enterococcus ecocci  600
#> 25 #CC3231 Enterococcus ecocci 3500
#> 26 #EE7600 Enterococcus ecocci  320
#> 27 #EE7600 Enterococcus ecocci  300
#> 28 #CC3231 Enterococcus ecocci 6100
#> 29 #EE7600 Enterococcus ecocci  580
#> 30 #CC3231 Enterococcus ecocci 1350
#> 31 #EE7600 Enterococcus ecocci  525
#> 32 #E9C318 Enterococcus ecocci   85
#> 33 #EE7600 Enterococcus ecocci  703
#> 34 #2DC938 Enterococcus ecocci   10
#> 35 #CC3231 Enterococcus ecocci 1990
#> 36 #EE7600 Enterococcus ecocci  171
#> 37 #EE7600 Enterococcus ecocci  717
#> 38 #EE7600 Enterococcus ecocci  727
#> 39 #EE7600 Enterococcus ecocci  908

# differentiate wet/dry samples in that time frame
anlz_enteromap(enterodata, yrsel = 2020, mosel = 9, wetdry = TRUE,
               temporal_window = 2, wet_threshold = 0.5)
#>                       station   yr mo Latitude Longitude ecocci       cat
#> 1          21FLCOSP_WQX-32-03 2020  9 27.80869 -82.61622     20      < 35
#> 2          21FLCOSP_WQX-44-02 2020  9 27.79258 -82.62640     41  35 - 129
#> 3          21FLCOSP_WQX-48-03 2020  9 27.71173 -82.68386     31      < 35
#> 4  21FLCOSP_WQX-CENTRAL CANAL 2020  9 27.72590 -82.68354     10      < 35
#> 5        21FLCOSP_WQX-COSP580 2020  9 27.73746 -82.69376     41  35 - 129
#> 6    21FLCOSP_WQX-NORTH CANAL 2020  9 27.72983 -82.68853     41  35 - 129
#> 7          21FLCOSP_WQX-SC-01 2020  9 27.75198 -82.63928    480 130 - 999
#> 8    21FLCOSP_WQX-SOUTH CANAL 2020  9 27.72461 -82.68794    360 130 - 999
#> 9      21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35
#> 10     21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35
#> 11     21FLDOH_WQX-MANATEE152 2020  9 27.53593 -82.73678     10      < 35
#> 12           21FLHILL_WQX-101 2020  9 28.02571 -82.63164    380 130 - 999
#> 13           21FLHILL_WQX-102 2020  9 28.01060 -82.60780    380 130 - 999
#> 14           21FLHILL_WQX-103 2020  9 27.99741 -82.58651    490 130 - 999
#> 15           21FLHILL_WQX-104 2020  9 27.98187 -82.56260    725 130 - 999
#> 16           21FLHILL_WQX-109 2020  9 27.94581 -82.40167      1      < 35
#> 17           21FLHILL_WQX-112 2020  9 27.70443 -82.44828    220 130 - 999
#> 18           21FLHILL_WQX-133 2020  9 27.91501 -82.40210   3100     > 999
#> 19           21FLHILL_WQX-144 2020  9 27.83802 -82.38191    620 130 - 999
#> 20           21FLHILL_WQX-180 2020  9 27.71599 -82.47044     56  35 - 129
#> 21             21FLHILL_WQX-2 2020  9 27.94183 -82.45854    280 130 - 999
#> 22            21FLHILL_WQX-22 2020  9 27.60810 -82.57120      2      < 35
#> 23           21FLHILL_WQX-264 2020  9 27.68253 -82.49677     40  35 - 129
#> 24           21FLHILL_WQX-560 2020  9 27.90522 -82.40163    600 130 - 999
#> 25           21FLHILL_WQX-561 2020  9 27.93162 -82.39926   3500     > 999
#> 26           21FLHILL_WQX-592 2020  9 27.77489 -82.38808    320 130 - 999
#> 27           21FLHILL_WQX-594 2020  9 27.81923 -82.38252    300 130 - 999
#> 28           21FLHILL_WQX-595 2020  9 27.82610 -82.38168   6100     > 999
#> 29           21FLHILL_WQX-596 2020  9 27.87821 -82.39036    580 130 - 999
#> 30           21FLHILL_WQX-597 2020  9 27.88509 -82.39170   1350     > 999
#> 31           21FLHILL_WQX-598 2020  9 27.89004 -82.39709    525 130 - 999
#> 32            21FLHILL_WQX-74 2020  9 27.85909 -82.38334     85  35 - 129
#> 33         21FLPDEM_WQX-12-02 2020  9 28.02048 -82.68955    703 130 - 999
#> 34         21FLPDEM_WQX-19-02 2020  9 27.92840 -82.74575     10      < 35
#> 35         21FLPDEM_WQX-22-01 2020  9 27.92346 -82.71765   1990     > 999
#> 36         21FLPDEM_WQX-24-02 2020  9 27.91209 -82.70025    171 130 - 999
#> 37         21FLPDEM_WQX-39-01 2020  9 27.75798 -82.73151    717 130 - 999
#> 38         21FLPDEM_WQX-39-05 2020  9 27.76021 -82.72979    727 130 - 999
#> 39       21FLTPA_WQX-G1SW0099 2020  9 27.86294 -82.61633    908 130 - 999
#>        col          ind  indnm conc wet_sample
#> 1  #2DC938 Enterococcus ecocci   20      FALSE
#> 2  #E9C318 Enterococcus ecocci   41      FALSE
#> 3  #2DC938 Enterococcus ecocci   31      FALSE
#> 4  #2DC938 Enterococcus ecocci   10      FALSE
#> 5  #E9C318 Enterococcus ecocci   41      FALSE
#> 6  #E9C318 Enterococcus ecocci   41      FALSE
#> 7  #EE7600 Enterococcus ecocci  480      FALSE
#> 8  #EE7600 Enterococcus ecocci  360      FALSE
#> 9  #2DC938 Enterococcus ecocci   10      FALSE
#> 10 #2DC938 Enterococcus ecocci   10      FALSE
#> 11 #2DC938 Enterococcus ecocci   10      FALSE
#> 12 #EE7600 Enterococcus ecocci  380      FALSE
#> 13 #EE7600 Enterococcus ecocci  380       TRUE
#> 14 #EE7600 Enterococcus ecocci  490       TRUE
#> 15 #EE7600 Enterococcus ecocci  725      FALSE
#> 16 #2DC938 Enterococcus ecocci    1      FALSE
#> 17 #EE7600 Enterococcus ecocci  220      FALSE
#> 18 #CC3231 Enterococcus ecocci 3100      FALSE
#> 19 #EE7600 Enterococcus ecocci  620      FALSE
#> 20 #E9C318 Enterococcus ecocci   56      FALSE
#> 21 #EE7600 Enterococcus ecocci  280       TRUE
#> 22 #2DC938 Enterococcus ecocci    2      FALSE
#> 23 #E9C318 Enterococcus ecocci   40      FALSE
#> 24 #EE7600 Enterococcus ecocci  600      FALSE
#> 25 #CC3231 Enterococcus ecocci 3500      FALSE
#> 26 #EE7600 Enterococcus ecocci  320      FALSE
#> 27 #EE7600 Enterococcus ecocci  300      FALSE
#> 28 #CC3231 Enterococcus ecocci 6100      FALSE
#> 29 #EE7600 Enterococcus ecocci  580      FALSE
#> 30 #CC3231 Enterococcus ecocci 1350      FALSE
#> 31 #EE7600 Enterococcus ecocci  525      FALSE
#> 32 #E9C318 Enterococcus ecocci   85       TRUE
#> 33 #EE7600 Enterococcus ecocci  703      FALSE
#> 34 #2DC938 Enterococcus ecocci   10      FALSE
#> 35 #CC3231 Enterococcus ecocci 1990      FALSE
#> 36 #EE7600 Enterococcus ecocci  171      FALSE
#> 37 #EE7600 Enterococcus ecocci  717      FALSE
#> 38 #EE7600 Enterococcus ecocci  727      FALSE
#> 39 #EE7600 Enterococcus ecocci  908      FALSE