Assign threshold categories to Fecal Indicator Bacteria (FIB) data

Assign threshold categories to Fecal Indicator Bacteria (FIB) data

Usage

anlz_fibcat(fibdata, yrsel = NULL, mosel = NULL, areasel = NULL)

Arguments

fibdata

input FIB data.frame as returned by read_importfib

yrsel

optional numeric value to filter output by years in fibdata

mosel

optional numeric value to filter output by month in fibdata

areasel

optional character string to filter output by stations in the "Alafia" or "Hillsborough" River basins, see details

Value

A data.frame similar to fibdata with additional columns describing station categories and optionally filtered by arguments passed to the function

Details

Categories based on relevant thresholds are assigned to each observation. The categories are specific to E. coli or Enterococcus and are assigned based on the station class as freshwater (class as 1 or 3F) or marine (class as 2 or 3M), respectively. A station is categorized into one of four ranges defined by the thresholds as noted in the cat column of the output, with corresponding colors appropriate for each range as noted in the col column of the output.

The areasel argument can indicate either "Alafia" or "Hillsborough" for the corresponding river basins, where rows in fibdata are filtered based on the the selection. The Alafia River basin includes values in the area column of fibdata as "Alafia River" and "Alafia River Tributary". The Hillsborough River basin includes values in the area column of fibdat as "Hillsborough River", "Hillsborough River Tributary", "Lake Thonotosassa", "Lake Thonotosassa Tributary", and "Lake Roberta". Not all areas may be present based on the selection.

Examples

# assign categories to all
anlz_fibcat(fibdata)
#> # A tibble: 20,225 × 12
#>    area    epchc_station class    yr    mo Latitude Longitude ecoli ecocci ind  
#>    <chr>           <dbl> <chr> <dbl> <dbl>    <dbl>     <dbl> <dbl>  <dbl> <chr>
#>  1 Hillsb…             2 3M     2023     4     27.9     -82.5    NA      2 Ente…
#>  2 Alafia…            74 3M     2023     4     27.9     -82.4    NA      2 Ente…
#>  3 Hillsb…           105 3M     2023     4     28.0     -82.4    NA      8 Ente…
#>  4 Hillsb…           106 1      2023     4     28.1     -82.4     5      4 E. c…
#>  5 Lake T…           107 3F     2023     4     28.0     -82.3   110    110 E. c…
#>  6 Hillsb…           108 3F     2023     4     28.1     -82.2   103     72 E. c…
#>  7 Alafia…           111 3F     2023     4     27.9     -82.2    NA     NA E. c…
#>  8 Alafia…           114 3F     2023     4     27.9     -82.3   184    100 E. c…
#>  9 Alafia…           115 3F     2023     4     27.9     -82.1    57     57 E. c…
#> 10 Alafia…           116 3F     2023     4     27.9     -82.1    32     52 E. c…
#> # ℹ 20,215 more rows
#> # ℹ 2 more variables: cat <fct>, col <chr>

# filter by year, month, and area
anlz_fibcat(fibdata, yrsel = 2020, mosel = 7, areasel = 'Alafia')
#> # A tibble: 24 × 12
#>    area    epchc_station class    yr    mo Latitude Longitude ecoli ecocci ind  
#>    <chr>           <dbl> <chr> <dbl> <dbl>    <dbl>     <dbl> <dbl>  <dbl> <chr>
#>  1 Alafia…            74 3M     2020     7     27.9     -82.4    NA    168 Ente…
#>  2 Alafia…           111 3F     2020     7     27.9     -82.2    50   7000 E. c…
#>  3 Alafia…           114 3F     2020     7     27.9     -82.3   447   1560 E. c…
#>  4 Alafia…           115 3F     2020     7     27.9     -82.1    80    260 E. c…
#>  5 Alafia…           116 3F     2020     7     27.9     -82.1    30    367 E. c…
#>  6 Alafia…          1301 3M     2020     7     27.9     -82.4    NA     NA Ente…
#>  7 Alafia…          1303 3M     2020     7     27.9     -82.4    NA     NA Ente…
#>  8 Alafia…          1304 3M     2020     7     27.9     -82.4    NA     NA Ente…
#>  9 Alafia…          1306 3M     2020     7     27.9     -82.3    NA     NA Ente…
#> 10 Alafia…          1307 3M     2020     7     27.9     -82.3    NA     NA Ente…
#> # ℹ 14 more rows
#> # ℹ 2 more variables: cat <fct>, col <chr>