Identify Fecal Indicator Bacteria samples as coming from a 'wet' or 'dry' time period

Identify Fecal Indicator Bacteria samples as coming from a 'wet' or 'dry' time period

Usage

anlz_fibwetdry(fibdata, precipdata, temporal_window = 2, wet_threshold = 0.5)

Arguments

fibdata

input data frame

precipdata

input data frame as returned by read_importrain. columns should be: station, date (yyyy-mm-dd), rain (in inches). The object catchprecip has this data from 1995-2023 for select stations.

temporal_window

numeric, number of days precipitation should be summed over (1 = day of sample only; 2 = day of sample + day before; etc.)

wet_threshold

numeric, inches accumulated through the defined temporal window, above which a sample should be defined as being from a 'wet' time period

Value

a data frame; the original fibdata data frame with three additional columns. rain_sampleDay is the total rain (inches) on the day of sampling, rain_total is the total rain (inches) for the period of time defined by temporal_window, and wet_sample is logical, indicating whether the rainfall for that station's catchment exceeded the amount over the time period specified in args.

Details

This function allows the user to specify a threshold for declaring a sample to be taken after an important amount of rain over an important amount of days, and declaring it to be 'wet'. This is of interest because samples taken after significant precipitation (definitions of this vary, which is why the user can specify desired thresholds) are more likely to exceed relevant bacterial thresholds. Identifying samples as 'wet' or not allows for calculation of further indices for wet and dry subsets of samples.

Examples

entero_wetdry <- anlz_fibwetdry(enterodata, catchprecip)
head(entero_wetdry)
#>         date            station ecocci ecocci_censored ecocci_units Latitude
#> 1 2013-06-27 21FLCOSP_WQX-32-03    740           FALSE    cfu/100mL 27.80869
#> 2 2013-06-27 21FLCOSP_WQX-32-03    740           FALSE    cfu/100mL 27.80869
#> 3 2013-08-05 21FLCOSP_WQX-32-03     60           FALSE    cfu/100mL 27.80869
#> 4 2013-08-05 21FLCOSP_WQX-32-03     60           FALSE    cfu/100mL 27.80869
#> 5 2013-09-11 21FLCOSP_WQX-32-03     28           FALSE    cfu/100mL 27.80869
#> 6 2013-09-11 21FLCOSP_WQX-32-03     28           FALSE    cfu/100mL 27.80869
#>   Longitude     time time_zone MDL   yr mo rain_sampleDay rain_total wet_sample
#> 1 -82.61622 08:01:00       EST  20 2013  6     0.20445455 0.90536364       TRUE
#> 2 -82.61622 08:23:00       EST  20 2013  6     0.20445455 0.90536364       TRUE
#> 3 -82.61622 08:01:00       EST  20 2013  8     0.05845455 0.05954545      FALSE
#> 4 -82.61622 08:16:00       EST  20 2013  8     0.05845455 0.05954545      FALSE
#> 5 -82.61622 08:01:00       EST   2 2013  9     0.00000000 0.00000000      FALSE
#> 6 -82.61622 08:40:00       EST   2 2013  9     0.00000000 0.00000000      FALSE