Estimate tidal creek report card scores

Estimate tidal creek report card scores

anlz_tdlcrk(tidalcreeks, iwrraw, tidtrgs = NULL, yr = 2021)

Arguments

tidalcreeks
sf object for population of tidal creeks

iwrraw
FDEP impaired waters data base as data.frame

tidtrgs
optional data.frame for tidal creek nitrogen targets, defaults to tidaltargets

yr
numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date

Value

A data.frame with the report card scores for each creek, as prioritize, investigate, caution, monitor, or no data

Examples

anlz_tdlcrk(tidalcreeks, iwrraw, yr = 2021)

#> # A tibble: 613 x 10
#>       id wbid  JEI   name    class monitor caution investigate prioritize score 
#>    <int> <chr> <chr> <chr>   <chr>   <int>   <int>       <int>      <int> <chr> 
#>  1     1 2052  CC01  Rock C… 3M         10      NA          NA         NA Monit…
#>  2     2 1983B CC01  Rock C… 2          NA      NA          NA         NA No Da…
#>  3     3 1983B CC02  Oyster… 2          NA      NA          NA         NA No Da…
#>  4     4 2067  CC02  Oyster… 3M         10      NA          NA         NA Monit…
#>  5     5 1983B CC03  Buck C… 2          NA      NA          NA         NA No Da…
#>  6     6 2068  CC03  Buck C… 3M          9       1          NA         NA Monit…
#>  7     7 2078A CC04  Buck C… 2          NA      NA           1         NA Inves…
#>  8     8 2078B CC05  Coral … 2           2      NA          NA         NA Monit…
#>  9     9 2078A CC05  Coral … 2          NA      NA          NA         NA No Da…
#> 10    10 2065C CC06  Catfis… 2          NA      NA          NA         NA No Da…
#> # … with 603 more rows