Estimate tidal creek report card scores

Estimate tidal creek report card scores

Usage

anlz_tdlcrk(tidalcreeks, iwrraw, tidtrgs = NULL, yr = 2023)

Arguments

tidalcreeks

sf object for population of tidal creeks

iwrraw

FDEP impaired waters data base as data.frame

tidtrgs

optional data.frame for tidal creek nitrogen targets, defaults to tidaltargets

yr

numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date

Value

A data.frame with the report card scores for each creek, as prioritize, investigate, caution, monitor, or no data

Examples

anlz_tdlcrk(tidalcreeks, iwrraw, yr = 2023)
#> # A tibble: 615 × 10
#>       id wbid  JEI   name     class monitor caution investigate prioritize score
#>    <int> <chr> <chr> <chr>    <chr>   <dbl>   <dbl>       <dbl>      <dbl> <chr>
#>  1     1 1983B CC01  Rock Cr… 2          NA      NA          NA         NA No D…
#>  2     2 2052  CC01  Rock Cr… 3M         10      NA          NA         NA Moni…
#>  3     3 1983B CC02  Oyster … 2          NA      NA          NA         NA No D…
#>  4     4 2067  CC02  Oyster … 3M         10      NA          NA         NA Moni…
#>  5     5 1983B CC03  Buck Cr… 2          NA      NA          NA         NA No D…
#>  6     6 2068  CC03  Buck Cr… 3M          9       1          NA         NA Moni…
#>  7     7 2078A CC04  Buck Cr… 2          NA      NA           1         NA Inve…
#>  8     8 2078A CC05  Coral C… 2          NA      NA          NA         NA No D…
#>  9     9 2078B CC05  Coral C… 2           4       1          NA         NA Moni…
#> 10    10 2065C CC06  Catfish… 2          NA      NA          NA         NA No D…
#> # ℹ 605 more rows