Analyze tidal creek water quality indicators

Estimate tidal creek water quality indicators to suppport report card scores

anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018)

Arguments

tidalcreeks
sf object for population of tidal creeks

iwrraw
FDEP impaired waters rule run 56 data base as data.frame

yr
numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date

Value

A data.frame with the indicator values for each tidal creek

Details

Annual geometric means for additional water quality data available at each wbid, JEI combination. Florida trophic state index values are also estimated where data are available. Nitrogen ratios are estimated for JEIs that cover source (upstream, freshwater) and tidal (downstream) WBIDs, defined as the ratio of concentrations between the two (i.e., ratios > 1 mean source has higher concentrations). Nitrogen ratios for a given year reflect the median of all daily ratios when concentrations were measured in both a source and tidal segment during the same day. Note that a ratio of one can be obtained if both the source and tidal segments are at minimum detection.

See also

Examples

dat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018)
head(dat)

#>   id  wbid  JEI         name class year CHLAC COLOR     COND       DO     DOSAT
#> 1  1  2052 CC01   Rock Creek    3M 2009    NA    NA 50923.64 5.998273  75.00467
#> 2  1  2052 CC01   Rock Creek    3M 2010    NA    NA 46485.21 8.849809 110.65252
#> 3  2 1983B CC01   Rock Creek     2 2009    NA    NA 52099.90 8.249848 103.14795
#> 4  2 1983B CC01   Rock Creek     2 2010    NA    NA 42588.26 8.348653 104.38271
#> 5  3 1983B CC02 Oyster Creek     2 2008     5    10 55530.00 5.134803  64.19720
#> 6  3 1983B CC02 Oyster Creek     2 2009    NA    NA 52161.41 7.239759  90.53120
#>   NO23 ORGN    SALIN  TKN   TN    TP TSS TURB chla_tn_ratio tn_tp_ratio
#> 1   NA   NA 33.25918   NA   NA    NA  NA   NA            NA          NA
#> 2   NA   NA 30.20121   NA   NA    NA  NA   NA            NA          NA
#> 3   NA   NA 34.34996   NA   NA    NA  NA   NA            NA          NA
#> 4   NA   NA 27.39106   NA   NA    NA  NA   NA            NA          NA
#> 5 0.03   NA 36.91500 1.11 1.14 0.094   8  5.2      4.385965    12.12766
#> 6   NA   NA 34.20891   NA   NA    NA  NA   NA            NA          NA
#>   chla_tsi   tn_tsi  tn2_tsi   tp_tsi  tp2_tsi  nut_tsi      tsi no23_ratio
#> 1       NA       NA       NA       NA       NA      NaN       NA         NA
#> 2       NA       NA       NA       NA       NA      NaN       NA         NA
#> 3       NA       NA       NA       NA       NA      NaN       NA         NA
#> 4       NA       NA       NA       NA       NA      NaN       NA         NA
#> 5 39.97591 58.59436 62.41899 66.10528 83.42176 62.34982 51.16286         NA
#> 6       NA       NA       NA       NA       NA      NaN       NA         NA