Get attainment categories for a chosen year

Get attainment categories for a chosen year

anlz_yrattain(epcdata, yrsel, partialyr = FALSE)

Arguments

epcdata
data frame of epc data returned by read_importwq

yrsel
numeric indicating chosen year

partialyr
logical indicating if incomplete annual data for the most recent year are approximated by five year monthly averages for each parameter

Value

A data.frame for the chosen year and all bay segments showing the bay segment averages for chloropyll concentration, light attenuations, segment targets, and attainment categories.

See also

Examples

# defaults to current year
anlz_yrattain(epcdata, yrsel = 2019)

#> # A tibble: 4 x 6
#>   bay_segment chla_val chla_target la_val la_target outcome
#>   <fct>          <dbl>       <dbl>  <dbl>     <dbl> <chr>  
#> 1 OTB             9.78         8.5  0.740      0.83 yellow 
#> 2 HB             11.0         13.2  0.950      1.58 green  
#> 3 MTB             5.69         7.4  0.561      0.83 green  
#> 4 LTB             3.91         4.6  0.619      0.63 green