Create a table for the annual management outcome assessments

Create a table for the annual management outcome assessments for chlorophyll-a and light attenuation by bay segment

Usage

show_annualassess(
  epcdata,
  yrsel,
  partialyr = F,
  caption = F,
  family = "Arial",
  txtsz = 12,
  width = NULL
)

Arguments

epcdata

data frame of epc data returned by read_importwq

yrsel

numeric indicating chosen year

partialyr

logical indicating if incomplete annual data for the most recent year are approximated by five year monthly averages for each parameter

caption

logical indicating if a caption is added using set_caption

family

chr string indicating font family for text labels

txtsz

numeric indicating font size

width

optional numeric value indicating width in inches

Value

A flextable object showing the segment-averaged chlorophyll-a and light attenuation for the selected year, with bay segment names colored by the management outcome used in show_matrix.

Examples

show_annualassess(epcdata, yrsel = 2023)

Segment

Chl-a (ug/L)

Light Penetration (m-1)

2023

target

2023

target

OTB

6.2

8.5

0.73

0.83

HB

6.9

13.2

0.79

1.58

MTB

3.7

7.4

0.47

0.83

LTB

2.6

4.6

0.54

0.63

show_annualassess(epcdata, yrsel = 2023, caption = TRUE)

Water quality outcomes for 2023.

Segment

Chl-a (ug/L)

Light Penetration (m-1)

2023

target

2023

target

OTB

6.2

8.5

0.73

0.83

HB

6.9

13.2

0.79

1.58

MTB

3.7

7.4

0.47

0.83

LTB

2.6

4.6

0.54

0.63