Plotly barplots of tidal creek context indicators

Plotly barplots of tidal creek context indicators

Usage

show_tdlcrkindic(
  id,
  cntdat,
  yr = 2022,
  thrsel = FALSE,
  pal = c("#5C4A42", "#427355", "#004F7E")
)

Arguments

id

numeric indicating the id number of the tidal creek to plot

cntdat

output from anlz_tdlcrkindic

yr

numeric indicating reference year

thrsel

logical if threshold lines and annotations are shown on the plots

pal

vector of colors for the palette

Value

A plotly object

Examples

cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2022)

set.seed(123)
id <- sample(unique(cntdat$id), 1)
show_tdlcrkindic(id, cntdat, thrsel = TRUE)