Plotly empirical CDF plots of tidal creek context indicators

Plotly empirical CDF plots of tidal creek context indicators

Usage

show_tdlcrkindiccdf(
  id,
  cntdat,
  yr = 2023,
  thrsel = FALSE,
  pal = c("#5C4A42", "#427355", "#004F7E")
)

Arguments

id

numeric indicating the id number of the tidal creek to plot

cntdat

output from anlz_tdlcrkindic

yr

numeric indicating reference year

thrsel

logical if threshold lines and annotations are shown on the plots

pal

vector of colors for the palette

Value

A plotly object

Details

This function returns several empirical cumulative distribution plots for the tidal creek context indicators. Points on the plot indicate the observed values and percentiles for the creek specified by id. The percentiles and CDF values are defined by the "population" of creeks in cntdat. Points in the plots are color-coded by sample year to evaluate temporal trends, if any.

Examples

cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2023)

set.seed(123)
id <- sample(unique(cntdat$id), 1)
show_tdlcrkindiccdf(id, cntdat, thrsel = TRUE)